\name{check_phylo4}
\alias{check_phylo4}
\title{Validity checking for phylo4 objects}
\description{
  Basic checks on the validity of S4 phylogenetic objects
}
\usage{
check_phylo4(object)
}
\arguments{
  \item{object}{A prospective S4 object}
}
\value{
  As required by \code{\link{validObject}}, returns an
  error string (describing problems) or TRUE if everything is OK
}
\author{Ben Bolker}
\keyword{misc}
