\name{printphylo}
\alias{printphylo}
\title{Deprecated functions}
\description{
  Functions which are no longer used (or suggested)
}
\usage{
printphylo(x, printlen = 6, ...)
}
\arguments{
  \item{x}{A phylo4 object}
  \item{printlen}{Maximum number of tip labels to print}
  \item{\dots}{Included for consistency}
}
\value{
  \code{printphylo} prints a short description of the object
}
\author{Ben Bolker (copied from print.phylo in ape package)}
\keyword{misc}
