\name{identify-methods}
\docType{methods}
\alias{identify-methods}
\alias{identify,phylo4-method}
\title{Graphically identify nodes in a phylogeny}
\description{
  Plots a phylogeny and allows the user to click
  on internal nodes to identify them
}
\section{Methods}{
\describe{
  \item{x = "phylo4"}{an object of class \code{phylo4}
    (or extensions)}
}}
\usage{
    \S4method{identify}{phylo4}(x,n=1,\dots)
}
\arguments{
  \item{x}{a phylo4 object}
  \item{n}{number of nodes to identify}
  \item{\dots}{additional arguments (for generic compatibility)}
  }
\keyword{methods}
