\name{owls4}
\alias{owls4}
\docType{data}
\title{
'Owls' data from ape
}
\description{
  A tiny tree, for testing/example purposes, using one of the examples from
the \code{ape} package
}
\usage{data(owls4)}
\format{
This is the standard 'owls' tree from the \code{ape} package,
in \code{phylo4} format.
}
\source{
From various examples in the \code{ape} package
}
\examples{
data(owls4)
plot(owls4)
}
\keyword{datasets}
