% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdata.R
\docType{class}
\name{pdata-class}
\alias{[,pdata,ANY,ANY,ANY-method}
\alias{[,pdata-method}
\alias{[<-,pdata-method}
\alias{[[,pdata,ANY,ANY-method}
\alias{[[,pdata,ANY,missing-method}
\alias{[[,pdata-method}
\alias{[[<-,pdata-method}
\alias{pdata-class}
\alias{ptypes}
\title{Class "pdata"}
\description{
Data class for phylo4d objects
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("pdata", ...)}.
}
\author{
Ben Bolker
}
\keyword{classes}

