\name{phylobase-package}
\alias{phylobase-package}
\alias{phylobase}
\docType{package}
\title{Utilities and Tools for Phylogenetics}
\description{
   Base package for phylogenetic structures and comparative data.
}

\details{
    \tabular{ll}{
        Package:    \tab phylobase\cr
        Type:       \tab Package\cr
        Date:       \tab 2009\cr
        Depends:    \tab methods, grid, ape(>= 2.1)\cr
        Suggests:   \tab ade4, MASS, gridBase\cr
        License:    \tab GPL Version 2 or later\cr
        Authors:    \tab R Hackathon (Ben Bolker, Marguerite Butler, Peter Cowan,  Damien de Vienne, Thibaut Jombart, Steve Kembel, Francois Michonneau, David Orme, Brian O'Meara, Emmanuel Paradis, Derrick Zwickl) \cr        
        URL:        \tab \url{http://phylobase.r-forge.r-project.org/}
    }
}

\section{MoreInfo}{
See the help index and run \code{vignette("phylobase", "phylobase")} -- also available at \url{../inst/doc/phylobase.pdf} for further details.
}

\keyword{package}
