\name{phylobubbles}
\alias{phylobubbles}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Bubble plots for phylo4d objects }
\description{
Plots either circles or squares corresponding to the magnitude of each cell of a \code{phylo4d} object.}
\usage{
phylobubbles(XXYY, square = FALSE, grid = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{XXYY}{ A list generated by the \code{phyloXXYY} function }
  \item{square}{ A logical indicating whether squares should be plotted instead of circles }
  \item{grid}{ A logical indicating whether a grey grid should be plotted behind the bubbles }
}
\value{A list consisting of:
\item{max}{ The maximum data value } 
\item{min}{ The minimum data value } 
\item{has.na}{ logical indicating whether the data contains \code{NA}s } 
\item{bubscale}{ a numeric giving the radius of the largest bubble in inches}
}

\author{Peter Cowan \email{pdc@berkeley.edu}}

\seealso{ \code{\link{phyloXXYY}}, \code{treePlot} }
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\keyword{methods}