\name{printphylo}
\alias{printphylo}
\title{print a phylogeny}
\description{
  Prints a short description of a phylo4 object
}
\usage{
printphylo(x, printlen = 6, ...)
}
\arguments{
  \item{x}{A phylo4 object}
  \item{printlen}{Maximum number of tip labels to print}
  \item{\dots}{Included for consistency}
}
\value{
  Prints a short description of the object
}
\author{Ben Bolker (copied from print.phylo in ape package)}
\keyword{misc}
