\name{read.nexustreestring}
\alias{read.nexustreestring}
\title{Read Nexus tree string}
\description{
  This is a stub!
}
\usage{
read.nexustreestring(X)
}
\arguments{
  \item{X}{a string containing a Nexus tree definition}
}
\value{
  Some kind of tree list?
}
\author{Brian O'Meara?}
\keyword{misc}
