\name{segs}
\alias{segs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Calculate the line segment vectors }
\description{
Calculates the line segments necessary to draw a phylogenetic tree.
}
\usage{
segs(XXYY)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{XXYY}{ A list generated by the \code{\line{phyloXXYY}} function }
}
\value{
A list containing:
  \item{ v0x }{ A vector with the beginning vertical line x values}
  \item{ v1x }{ A vector with the end vertical line x values}
  \item{ v0y }{ A vector with the beginning vertical line y values}
  \item{ v1y }{ A vector with the end vertical line y values}
  \item{ h0x }{ A vector with the beginning horizontal line x values}
  \item{ h1x }{ A vector with the end horizontal line x values}
  \item{ h0Y }{ A vector with the beginning horizontal line y values}
  \item{ h1Y }{ A vector with the end horizontal line y values}
}
\author{Peter Cowan \email{pdc@berkeley.edu}}

\seealso{ \code{\link{phyloXXYY}}, \code{treePlot}, \code{\link{tree.plot}} }

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}

\keyword{methods}
