\name{geospiza}
\alias{geospiza}
\docType{data}
\title{Data from Darwin's finches}
\description{
  Phylogenetic tree and morphological data for Darwin's finches
}
\usage{data(geospiza)}
\format{
  A \code{phylo4d} object
}
\note{
  Stolen from Luke Harmon's Geiger package,
  to avoid unnecessary dependencies
}
\source{
  Dolph Schluter via Luke Harmon
}
\examples{
data(geospiza)
plot(geospiza)
}
\keyword{datasets}
