% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylobase-package.R
\docType{data}
\name{geospiza}
\alias{geospiza}
\alias{geospiza_raw}
\title{Data from Darwin's finches}
\format{\code{geospiza} is a \code{phylo4d} object; \code{geospiza_raw} is a
list containing \code{tree}, a \code{phylo} object (the tree), \code{data},
and a data frame with the data (for showing examples of how to merge tree
and data)}
\source{
Dolph Schluter via Luke Harmon
}
\description{
Phylogenetic tree and morphological data for Darwin's finches, in different
formats
}
\note{
Stolen from Luke Harmon's Geiger package, to avoid unnecessary
dependencies
}
\examples{

data(geospiza)
plot(geospiza)

}
\keyword{datasets}
