% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiphylo4-class.R
\docType{class}
\name{multiPhylo-class}
\alias{multiPhylo-class}
\alias{multiPhylo4-class}
\alias{multiPhylo4d-class}
\alias{tbind}
\title{multiPhylo4 and extended classes}
\description{
Classes for lists of phylogenetic trees.  These classes and methods are
planned for a future version of \code{phylobase}.

Classes for lists of phylogenetic trees.  These classes and methods are
planned for a future version of \code{phylobase}.
}
\keyword{classes}
