% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylobase-package.R
\docType{data}
\name{owls4}
\alias{owls4}
\title{'Owls' data from ape}
\format{This is the standard 'owls' tree from the \code{ape} package, in
\code{phylo4} format.}
\source{
From various examples in the \code{ape} package
}
\description{
A tiny tree, for testing/example purposes, using one of the examples from
the \code{ape} package
}
\examples{

data(owls4)
plot(owls4)

}
\keyword{datasets}
