% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdata.R
\name{pdata}
\alias{pdata}
\alias{check_pdata}
\title{Constructor for pdata (phylogenetic data) class}
\usage{
pdata(data, type, comment, metadata)
}
\arguments{
\item{data}{a data frame}

\item{type}{a factor with levels as specified by \linkS4class{pdata}, the
same length as \code{ncol(data)}}

\item{comment}{a character vector, the same length as \code{ncol(data)}}

\item{metadata}{an arbitrary list}
}
\value{
An object of class \code{pdata}
}
\description{
Combine data, type, comments, and metadata information to create a new pdata
object, or check such an object for consistency
}
\seealso{
\linkS4class{pdata}
}
\author{
Ben Bolker
}
\keyword{misc}
