\docType{class}
\name{phylo4-class}
\alias{phylo4-class}
\title{The phylo4 class}
\description{
Classes for phylogenetic trees
}
\section{Objects from the Class}{
  Phylogenetic tree objects can be created by calls to the
  \code{\link{phylo4}} constructor function.  Translation
  functions from other phylogenetic packages are also
  available. See \code{\link{coerce-methods}}.
}
\author{
Ben Bolker, Thibaut Jombart
}
\seealso{
The \code{\link{phylo4}} constructor, the
\code{\link{checkPhylo4}} function to check the validity of
\code{phylo4} objects. See also the \code{\link{phylo4d}}
constructor and the \linkS4class{phylo4d} class.
}
\keyword{classes}

