% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo4d-class.R
\docType{class}
\name{phylo4d-class}
\alias{phylo4d-class}
\title{phylo4d class}
\description{
S4 class for phylogenetic tree and data.
}
\section{Objects from the Class}{
 Objects can be created from various trees
and a data.frame using the constructor \code{phylo4d}, or using
\code{new("phylo4d", \dots{})} for empty objects.
}

\examples{
  example(read.tree, "ape")
  obj <- phylo4d(as(tree.owls.bis,"phylo4"), data.frame(wing=1:3))
  obj
  names(obj)
  summary(obj)
}
\seealso{
\code{\link{coerce-methods}} for translation
functions. The \code{\link{phylo4d-methods}} constructor. See also
the \code{\link{phylo4-methods}} constructor, the
\linkS4class{phylo4} class, and the \code{\link{checkPhylo4}}
function to check the validity of \code{phylo4} trees.
}
\author{
Ben Bolker, Thibaut Jombart
}
\keyword{classes}
