\name{phylobase-package}
\alias{phylobase-package}
\alias{phylobase}
\docType{package}
\title{Utilities and Tools for Phylogenetics}
\description{
   Base package for phylogenetic structures and comparative data.
}

\details{
    \itemize{
        \item{Package:}{phylobase}
        \item{Type:}{Package}
        \item{Date:}{2009}
	\item{Depends:}{methods, grid, ape(>= 2.1)}
	\item{Suggests:}{adephylo, MASS}
	\item{License:}{GPL Version 2 or later}
        \item{Authors:}{R Hackathon et al. (alphabetically: Ben Bolker,
	  Marguerite Butler, Peter Cowan,  Damien de Vienne, Thibaut Jombart,
	  Steve Kembel, Francois Michonneau, David Orme, Brian O'Meara,
	  Emmanuel Paradis, Jim Regetz, Derrick Zwickl )}
	\item{URL:}{\url{http://phylobase.r-forge.r-project.org/}}
    }
}

\section{MoreInfo}{
  See the help index \code{help(package="phylobase")} and run
  \code{vignette("phylobase", "phylobase")} for further details and
  examples about how to use \code{phylobase}.
}

\keyword{package}
