% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylobase.options.R
\name{phylobase.options}
\alias{phylobase.options}
\title{Set or return options of phylobase}
\usage{
phylobase.options(...)
}
\arguments{
\item{\dots}{a list may be given as the only argument, or any
number of arguments may be in the \code{name=value} form, or no
argument at all may be given.  See the Value and Details sections
for explanation.}
}
\value{
A list with the updated values of the parameters. If
arguments are provided, the returned list is invisible.
}
\description{
Provides a mean to control the validity of \code{phylobase}
objects such as singletons, reticulated trees, polytomies, etc.
}
\details{
The parameter values set via a call to this function will remain
in effect for the rest of the session, affecting the subsequent
behavior of phylobase.
}
\examples{
\dontrun{
phylobase.options(poly="fail")
# subsequent trees with polytomies will fail the validity check
}

}
\author{
Francois Michonneau (adapted from the package \code{sm})
}
\keyword{phylobase}
\keyword{validator}
