\name{phylog-class}
\docType{class}
\alias{phylog-class}
\alias{coerce,phylo4,phylog-method}
\title{Class "phylog"}
\description{S4 version of the class \code{phylog} from the \code{ade4} package.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
}
\section{Methods}{
  \describe{
    \item{coerce}{\code{signature(from = "phylo4", to = "phylog")}:
      a \linkS4class{phylo4} object can be coerced to a \code{phylog} object using \code{as(object,"phylog")}. }
	 }
}
\author{Thibaut Jombart \email{jombart@biomserv.univ-lyon1.fr}}
\seealso{The original \code{\link[ade4]{phylog}} from the
  \code{ade4} package. 
}
% commented to eliminate Note: on build
%\examples{
%}
\keyword{classes}
