\name{treewalk}
\alias{getAncest}
\alias{getDesc}
\title{Stubs for tree traversal etc.}
\description{
  Eventually these will be functions for moving around trees
}
\usage{
getAncest(phy, node)
}
\arguments{
  \item{phy}{a \code{phylo4} object}
  \item{node}{a node number (or name?)}
}
\details{
  Stubs.
}
\value{
  a node number or name
}
\keyword{misc}
